back-end/api/requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
six==1.16.0
Werkzeug==2.2.3
